requirements.txt
git+https://github.com/PierreMarchand20/asciinema_automation.git@main#egg=asciinema-automation
pyyaml==6.0
pydantic==1.10.5
